requirements.txt
asgiref==3.2.3
Django==3.0.1
django-modeltranslation==0.14.1
gunicorn==20.0.4
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
whitenoise==5.0.1
